refactors UserService, adds ProjectUserService
