+++ service/src/main/java/com/tu/FinancialQuickCheck/Service/UserService.java
@@ -2,7 +2,6 @@ package com.tu.FinancialQuickCheck.Service;
 
 import com.tu.FinancialQuickCheck.Exceptions.ResourceNotFound;
 import com.tu.FinancialQuickCheck.db.*;
-import com.tu.FinancialQuickCheck.dto.ProjectUserDto;
 import com.tu.FinancialQuickCheck.dto.UserDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,17 +20,11 @@ import java.util.UUID;
 @Service
 public class UserService {
 
-    private final UserRepository userRepository;
-    private final ProjectRepository projectRepository;
-    private final ProjectUserRepository projectUserRepository;
+    private final UserRepository repository;
 
     @Autowired
-    public UserService(UserRepository userRepository,
-                       ProjectRepository projectRepository,
-                       ProjectUserRepository projectUserRepository) {
-        this.userRepository = userRepository;
-        this.projectRepository = projectRepository;
-        this.projectUserRepository = projectUserRepository;
+    public UserService(UserRepository userRepository) {
+        this.repository = userRepository;
     }
 
     /**
@@ -49,7 +42,7 @@ public class UserService {
         newUser.username = userDto.username;
         newUser.email = userDto.email;
         newUser.password = userDto.password;
-        userRepository.save(newUser);
+        repository.save(newUser);
         return new UserDto(UUID.fromString(newUser.id), newUser.email, newUser.username);
     }
 
@@ -60,7 +53,7 @@ public class UserService {
     public List<UserDto> findAllUser() {
 
         List<UserDto> userList = new ArrayList<>();
-        Iterable<UserEntity> allUserEntitys = userRepository.findAll();
+        Iterable<UserEntity> allUserEntitys = repository.findAll();
 
         for (UserEntity userEntity : allUserEntitys) {
             UserDto userDto = new UserDto(UUID.fromString(userEntity.id), userEntity.email, userEntity.username);
@@ -81,7 +74,7 @@ public class UserService {
      */
     public UserDto findByEmail(String email) {
 
-        Iterable<UserEntity> allUserEntitys = userRepository.findAll();
+        Iterable<UserEntity> allUserEntitys = repository.findAll();
 
         for (UserEntity userEntity : allUserEntitys) {
             if (userEntity.email.equals(email)) {
@@ -102,7 +95,7 @@ public class UserService {
      */
     public void updateByEmail(UserDto userDto, String email) {
 
-        Iterable<UserEntity> allUserEntitys = userRepository.findAll();
+        Iterable<UserEntity> allUserEntitys = repository.findAll();
 
         for (UserEntity userEntity : allUserEntitys) {
             if (userEntity.email.equals(email)) {
@@ -117,7 +110,7 @@ public class UserService {
                 if (userDto.username != null) {
                     userEntity.username = userDto.username;
                 }
-                userRepository.save(userEntity);
+                repository.save(userEntity);
                 return;
             }
         }
@@ -129,9 +122,9 @@ public class UserService {
 
         UUID u = UUID.fromString(userID);
 
-        Optional<UserEntity> entity = userRepository.findById(userID);
+        Optional<UserEntity> entity = repository.findById(userID);
 
-        if (!userRepository.existsById(userID)) {
+        if (!repository.existsById(userID)) {
             throw new ResourceNotFound("userID " + userID + " not found");
         } else {
 
@@ -149,7 +142,7 @@ public class UserService {
                             user.username = userDto.username;
                         }
 
-                        return userRepository.save(user);
+                        return repository.save(user);
                     });
 
         }
@@ -164,15 +157,15 @@ public class UserService {
 
     public void deleteUser(String email) {
 
-        Iterable<UserEntity> allUserEntitys = userRepository.findAll();
+        Iterable<UserEntity> allUserEntitys = repository.findAll();
 
         for (UserEntity userEntity : allUserEntitys) {
             if (userEntity.email.equals(email)) {
-                Optional<UserEntity> tmpEnt = userRepository.findById(userEntity.id);
+                Optional<UserEntity> tmpEnt = repository.findById(userEntity.id);
                 if (tmpEnt.isEmpty()) {
                     throw new ResourceNotFound("User email |" + email + "| not found");
                 } else {
-                    userRepository.deleteById(tmpEnt.get().id);
+                    repository.deleteById(tmpEnt.get().id);
                 }
                 return;
             }
@@ -181,79 +174,13 @@ public class UserService {
 
 
     public void deleteUserById(String userID) {
-        Optional<UserEntity> entity = userRepository.findById(userID);
+        Optional<UserEntity> entity = repository.findById(userID);
 
         if (entity.isEmpty()) {
             throw new ResourceNotFound("userID " + userID + " not found");
         } else {
-            userRepository.deleteById(userID);
+            repository.deleteById(userID);
         }
 
-    }
-
-
-    public List<ProjectUserDto> getProjectUsersByProjectId(int projectID){
-
-        Optional<ProjectEntity> entity = projectRepository.findById(projectID);
-
-        if (entity.isEmpty()) {
-            throw new ResourceNotFound("projectID " + projectID + " not found");
-        } else {
-
-            Iterable<ProjectUserEntity> entities = entity.get().projectUserEntities;
-
-            List<ProjectUserDto> projectUserDtos = new ArrayList<>() {
-            };
-
-            for (ProjectUserEntity tmp : entities) {
-                projectUserDtos.add(new ProjectUserDto(
-                        UUID.fromString(tmp.projectUserId.getUserid().id),
-                        tmp.role,
-                        tmp.projectUserId.getUserid().email,
-                        tmp.projectUserId.getProjectid().id,
-                        tmp.projectUserId.getUserid().username
-                ));
-            }
-
-            return projectUserDtos;
-        }
-    }
-
-
-    public void createProjectUser(int projectID, String userID, ProjectUserDto projectUserDto){
-
-        ProjectUserEntity entity = new ProjectUserEntity();
-
-        if(!userRepository.existsById(userID) || !projectRepository.existsById(projectID)){
-            throw new ResourceNotFound("User or project does not exist.");
-        }else{
-            entity.projectUserId = new ProjectUserId(
-                    projectRepository.findById(projectID).get(),
-                    userRepository.findById(userID).get()
-                    );
-            entity.role = projectUserDto.role;
-
-            projectUserRepository.save(entity);
-        }
-    }
-
-
-    public void updateProjectUser(int projectID, UUID userID, ProjectUserDto projectUserDto){
-
-        if(!projectUserRepository.existsById(new ProjectUserId(projectRepository.findById(projectID).get(),
-                userRepository.findById(userID.toString()).get()))){
-            throw new ResourceNotFound("User is not assigned to project.");
-        }else{
-            projectUserRepository.findById(new ProjectUserId(projectRepository.findById(projectID).get(),
-                    userRepository.findById(userID.toString()).get()))
-                    .map(
-                        projectUser -> {
-                            if(projectUserDto.role != null){
-                                projectUser.role = projectUserDto.role;
-                            }
-
-                            return projectUserRepository.save(projectUser);
-                        });
-        }
     }
 }
